Create irat app and refactoring question app
