+++ pgtbl/irat/models/__init__.py
@@ -1,0 +1,1 @@
+from .irat_submission import IRATSubmission
