+++ pgtbl/irat/static/irat/css/list.css
@@ -1,0 +1,104 @@
+.panel-heading {
+  background-color: #286090 !important;
+  color: white !important;
+}
+
+.questions {
+  border-radius: 5px 5px 0 0;
+  margin-bottom: 0;
+}
+
+.course {
+  padding-right: 34px;
+}
+
+.classes {
+  padding-right: 5px;
+}
+
+.fa-book {
+  padding-right: 7px;
+}
+
+.teacher {
+  padding-right: 4px;
+}
+
+.fa-user {
+  padding-right: 11px;
+}
+
+.roles {
+  cursor: pointer;
+  color: whitesmoke;
+  font-size: 25px;
+}
+
+.roles:hover, .roles:focus {
+  color: white;
+}
+
+.progress-bar {
+  background-color: #286090;
+}
+
+/* Modals */
+.modal-footer {
+  background-color: #263238 !important;
+  border-bottom-left-radius: 10px;
+  border-bottom-right-radius: 10px;
+}
+
+.modal-body {
+  color: black;
+}
+
+/* Add button */
+.add-button {
+  border-radius: 50%;
+  position: fixed;
+  bottom: 10px;
+  right: 30px;
+}
+
+.btn:active:focus, .btn:focus {
+  outline: none;
+}
+
+.submit-button {
+  width: 100%;
+  margin-bottom: 10px;
+}
+
+/* iRAT */
+.time {
+  position:relative;
+  bottom: 25px;
+  left: 25px;
+}
+
+.time > a {
+  color: white;
+}
+
+.time > a:hover {
+  text-decoration: none;
+}
+
+.date {
+  position: relative;
+  top: 20px;
+  left: 25px;
+}
+
+.date > a {
+  color: white;
+}
+
+.date > a:hover {
+  text-decoration: none;
+}
+
+.form-group {
+  padding-top: 12px;
+}
