+++ pgtbl/irat/static/irat/css/result.css
@@ -1,0 +1,32 @@
+.questions {
+  border-radius: 5px 5px 0 0;
+  margin-bottom: 0;
+}
+
+.course {
+  padding-right: 34px;
+}
+
+.classes {
+  padding-right: 5px;
+}
+
+.fa-book {
+  padding-right: 7px;
+}
+
+.teacher {
+  padding-right: 4px;
+}
+
+.fa-user {
+  padding-right: 11px;
+}
+
+.progress-bar {
+  background-color: #286090;
+}
+
+.reset {
+  margin-top: 24px;
+}
