+++ pgtbl/irat/templates/irat/pagination.html
@@ -22,7 +22,7 @@
      {% trans 'Next' %}
   </a>
 {% else %}
-  <a href="{% url 'questions:irat-result' discipline.slug session.id %}"
+  <a href="{% url 'irat:result' discipline.slug session.id %}"
      class="btn btn-primary pull-right">
      <i class="fa fa-flag-checkered"></i>
      {% trans 'Conclude' %}
