+++ pgtbl/irat/views/__init__.py
@@ -1,0 +1,5 @@
+from .irat_answer_question_view import IRATAnswerQuestionView
+from .irat_date_update_view import IRATDateUpdateView
+from .irat_result_view import IRATResultView
+from .irat_update_view import IRATUpdateView
+from .irat_view import IRATView
