+++ pgtbl/irat/views/irat_update_view.py
@@ -1,27 +1,25 @@
 from django.contrib.auth.mixins import LoginRequiredMixin
-from django.shortcuts import redirect
 from django.utils.translation import ugettext_lazy as _
 from django.urls import reverse_lazy
 from django.contrib import messages
-from django.utils import timezone
 from django.views.generic import UpdateView
 
 from core.permissions import PermissionMixin
 from disciplines.models import Discipline
 from modules.models import TBLSession
-from questions.forms import IRATDateForm
+from irat.forms import IRATForm
 
 
-class IRATDateUpdateView(LoginRequiredMixin,
-                         PermissionMixin,
-                         UpdateView):
+class IRATUpdateView(LoginRequiredMixin,
+                     PermissionMixin,
+                     UpdateView):
     """
-    Update the iRAT date.
+    Update the iRAT duration and weight
     """
 
     model = TBLSession
     template_name = 'irat/irat.html'
-    form_class = IRATDateForm
+    form_class = IRATForm
 
     # Permissions
     permissions_required = ['crud_tests']
@@ -42,29 +40,9 @@ class IRATDateUpdateView(LoginRequiredMixin,
         Return the form with fields valided.
         """
 
-        now = timezone.localtime(timezone.now())
+        messages.success(self.request, _('iRAT updated successfully.'))
 
-        if form.instance.irat_datetime is None:
-
-            messages.error(
-                self.request,
-                _("iRAT date must to be filled in.")
-            )
-
-            return redirect(self.get_success_url())
-
-        if now > form.instance.irat_datetime:
-
-            messages.error(
-                self.request,
-                _("iRAT date must to be later than today's date.")
-            )
-
-            return redirect(self.get_success_url())
-
-        messages.success(self.request, _('iRAT date updated successfully.'))
-
-        return super(IRATDateUpdateView, self).form_valid(form)
+        return super(IRATUpdateView, self).form_valid(form)
 
     def get_success_url(self):
         """
@@ -72,7 +50,7 @@ class IRATDateUpdateView(LoginRequiredMixin,
         """
 
         success_url = reverse_lazy(
-            'questions:irat-list',
+            'irat:list',
             kwargs={
                 'slug': self.kwargs.get('slug', ''),
                 'pk': self.kwargs.get('pk', '')
