+++ pgtbl/modules/templates/modules/sidebar.html
@@ -46,7 +46,7 @@
     </a>
     <!-- Session iRAT menu -->
     {% if date > irat_datetime and user in discipline.students.all %}
-      <a href="{% url 'questions:irat-result' discipline.slug session.id %}"
+      <a href="{% url 'irat:result' discipline.slug session.id %}"
          class="btn"
          data-toggle="tooltip"
          data-placement="right"
@@ -54,7 +54,7 @@
          <i class="fa fa-street-view icon-lg"></i>
       </a>
     {% else %}
-      <a href="{% url 'questions:irat-list' discipline.slug session.id %}"
+      <a href="{% url 'irat:list' discipline.slug session.id %}"
          class="btn"
          data-toggle="tooltip"
          data-placement="right"
