+++ pgtbl/questions/forms/__init__.py
@@ -1,4 +1,3 @@
 from .alternative_form import AlternativeForm, AlternativeFormSet
 from .question_form import QuestionForm, AnswerQuestionForm
-from .irat_form import IRATForm, IRATDateForm
 from .grat_form import GRATForm, GRATDateForm, AnswerGRATQuestionForm
