+++ pgtbl/questions/models/__init__.py
@@ -1,4 +1,3 @@
 from .alternative import Alternative
 from .question import Question
-from .irat_submission import IRATSubmission
 from .grat_submission import GRATSubmission
