+++ pgtbl/questions/views/question/delete_question_view.py
@@ -76,7 +76,7 @@ class DeleteQuestionView(LoginRequiredMixin,
             )
         else:
             success_url = reverse_lazy(
-                'questions:irat-list',
+                'irat:list',
                 kwargs={
                     'slug': discipline.slug,
                     'pk': session.id
