+++ pgtbl/tbl/urls.py
@@ -16,6 +16,7 @@ urlpatterns = [
     url(r'', include('peer_review.urls')),
     url(r'', include('practical_test.urls')),
     url(r'', include('exercises.urls')),
+    url(r'', include('irat.urls')),
 ]
 
 # While in development mode we will use relative URL for static and average
